Revert "New Feature: Soft Delete Goal"
