+++ .eslintrc.json
@@ -21,7 +21,6 @@
   "extends": [
     "airbnb",
     "eslint:recommended",
-    "@typescript-eslint/parser",
     "plugin:react/recommended",
     "plugin:import/errors",
     "plugin:react/recommended",
