+++ playwright/utils/collaboration-feature-utils.ts
@@ -14,7 +14,7 @@ export async function goalActionFlow(page: Page, action: string) {
     .locator("div")
     .filter({ hasText: new RegExp(`^${action}$`) })
     .first()
-    .click({ force: true });
+    .click();
 }
 
 export async function goToShareGoalModalFlow(page: Page) {
