+++ src/components/GoalsComponents/DisplayChangesModal/DisplayChangesModal.tsx
@@ -15,7 +15,6 @@ import { archiveUserGoal, getGoal, removeGoalWithChildrens, updateGoal } from "@
 import { deleteGoalChangesInID, getInboxItem, removeGoalInbox, removePPTFromInboxOfGoal } from "@src/api/InboxAPI";
 import { findGoalTagChanges, jumpToLowestChanges } from "@src/helpers/GoalProcessor";
 import { acceptSelectedSubgoals, acceptSelectedTags } from "@src/helpers/InboxProcessor";
-import { getDeletedGoal, restoreUserGoal } from "@src/api/TrashAPI";
 import SubHeader from "@src/common/SubHeader";
 import ContactItem from "@src/models/ContactItem";
 import ZModal from "@src/common/ZModal";
@@ -164,8 +163,6 @@ const DisplayChangesModal = () => {
       await removeGoalWithChildrens(activeGoal);
     } else if (currentDisplay === "archived") {
       await archiveUserGoal(activeGoal);
-    } else if (currentDisplay === "restored") {
-      await restoreUserGoal(activeGoal);
     }
     setCurrentDisplay("none");
   };
@@ -188,28 +185,20 @@ const DisplayChangesModal = () => {
         const currPPT = participants[activePPT].relId;
         setActivePPT(0);
         setParticipants([...participants.filter((ele) => ele.relId !== currPPT)]);
-      } else if (typeAtPriority === "restored") {
-        const goalToBeRestored = await getDeletedGoal(parentId);
-        console.log("🚀 ~ getChanges ~ goalToBeRestored:", goalToBeRestored);
-        if (goalToBeRestored) {
-          delete goalToBeRestored.deletedAt;
-          setActiveGoal(goalToBeRestored);
-        }
-      } else {
-        const changedGoal = await getGoal(parentId);
-        if (changedGoal) {
-          setActiveGoal({ ...changedGoal });
-          if (typeAtPriority === "subgoals") {
-            setNewGoals(goals || []);
-          } else if (typeAtPriority === "modifiedGoals") {
-            setUpdatesIntent(goals[0].intent);
-            const incGoal: GoalItem = { ...goals[0].goal };
-            setUpdateList({ ...findGoalTagChanges(changedGoal, incGoal) });
-          }
-        }
       }
-      if (currentDisplay !== typeAtPriority) {
-        setCurrentDisplay(typeAtPriority);
+      const changedGoal = await getGoal(parentId);
+      if (changedGoal) {
+        setActiveGoal({ ...changedGoal });
+        if (typeAtPriority === "subgoals") {
+          setNewGoals(goals || []);
+        } else if (typeAtPriority === "modifiedGoals") {
+          setUpdatesIntent(goals[0].intent);
+          const incGoal: GoalItem = { ...goals[0].goal };
+          setUpdateList({ ...findGoalTagChanges(changedGoal, incGoal) });
+        }
+        if (currentDisplay !== typeAtPriority) {
+          setCurrentDisplay(typeAtPriority);
+        }
       }
     }
   };
@@ -284,10 +273,9 @@ const DisplayChangesModal = () => {
               />
             </p>
           )}
-          {["deleted", "archived", "restored"].includes(currentDisplay) && <div />}
+          {(currentDisplay === "archived" || currentDisplay === "deleted") && <div />}
           {currentDisplay === "modifiedGoals" && getEditChangesList()}
           {currentDisplay === "subgoals" && getSubgoalsList()}
-
           <div style={{ display: "flex", gap: 18, justifyContent: "flex-end" }}>
             {activeGoal && (
               <>
