+++ src/models/InboxItem.ts
@@ -1,6 +1,6 @@
 import { GoalItem } from "./GoalItem";
 
-export type typeOfChange = "subgoals" | "modifiedGoals" | "archived" | "deleted" | "restored";
+export type typeOfChange = "subgoals" | "modifiedGoals" | "archived" | "deleted";
 
 export type typeOfIntent = "suggestion" | "shared";
 export type changesInId = { level: number; id: string; intent: typeOfIntent };
@@ -11,7 +11,6 @@ export interface IChangesInGoal {
   modifiedGoals: changesInGoal[];
   archived: changesInId[];
   deleted: changesInId[];
-  restored: changesInId[];
 }
 
 export interface InboxItem {
