--- src/models/TrashItem.ts
@@ -1,5 +1,0 @@
-import { GoalItem } from "./GoalItem";
-
-export interface TrashItem extends GoalItem {
-  deletedAt: String;
-}
