+++ src/models/db.ts
@@ -7,7 +7,6 @@ import { InboxItem } from "./InboxItem";
 import { TaskItem } from "./TaskItem";
 import { GCustomItem } from "./GCustomItem";
 import { DumpboxItem } from "./DumpboxItem";
-import { TrashItem } from "./TrashItem";
 
 export const dexieVersion = 17;
 
@@ -31,8 +30,6 @@ export class ZinZenDB extends Dexie {
 
   dumpboxCollection!: Table<DumpboxItem, string>;
 
-  goalTrashCollection!: Table<TrashItem, string>;
-
   constructor() {
     super("ZinZenDB");
     this.version(dexieVersion)
@@ -52,8 +49,6 @@ export class ZinZenDB extends Dexie {
         customizationCollection: "++id, goalId, posIndex",
         dumpboxCollection: "id, key, value",
         partnersCollection: null,
-        goalTrashCollection:
-          "id, deletedAt, title, duration, sublist, habit, on, start, due, afterTime, beforeTime, createdAt, parentGoalId, archived, participants, goalColor, language, link, rootGoalId, timeBudget, typeOfGoal",
       })
       .upgrade((trans) => {
         console.log("🚀 ~ file: db.ts:63 ~ ZinZenDB ~ .upgrade ~ this.verno:", currentVersion);
