+++ src/store/index.ts
@@ -3,38 +3,26 @@ import { confirmActionState } from "@src/Interfaces/IPopupModals";
 import ContactItem from "@src/models/ContactItem";
 import { darkModeState } from "./DarkModeState";
 import { languageSelectionState } from "./LanguageSelectionState";
-import { isJSONParsable } from "@src/utils/patterns";
-
-const defaultConfirmationObj: confirmActionState = {
-  open: false,
-  goal: {
-    archive: true,
-    delete: true,
-    shareAnonymously: true,
-    shareWithOne: true,
-    restore: true,
-  },
-  collaboration: {
-    colabRequest: true,
-    delete: true,
-    archive: true,
-    restore: true,
-  },
-};
-
-if (isJSONParsable(localStorage.getItem("confirmationState"))) {
-  //@ts-ignore
-  const savedConfirmationState = JSON.parse(localStorage.getItem("confirmationService"));
-  defaultConfirmationObj.goal = { ...defaultConfirmationObj.goal, ...(savedConfirmationState?.goal || {}) };
-  defaultConfirmationObj.collaboration = {
-    ...defaultConfirmationObj.collaboration,
-    ...(savedConfirmationState?.collaboration || {}),
-  };
-}
 
 export const displayConfirmation = atom({
   key: "displayConfirmation",
-  default: defaultConfirmationObj,
+  default: JSON.parse(
+    localStorage.getItem("confirmationState") ||
+      JSON.stringify({
+        open: false,
+        goal: {
+          archive: true,
+          delete: true,
+          shareAnonymously: true,
+          shareWithOne: true,
+        },
+        collaboration: {
+          colabRequest: true,
+          delete: true,
+          archive: true,
+        },
+      }),
+  ) as confirmActionState,
 });
 
 export const lastAction = atom({
