+++ src/utils/defaultGenerators.ts
@@ -27,6 +27,5 @@ export function getDefaultValueOfGoalChanges() {
     modifiedGoals: [],
     archived: [],
     deleted: [],
-    restored: [],
   };
 }
